add keywordList

+++ DS_G15/src/Keyword.java
@@ -1,9 +1,9 @@
 
 public class Keyword {
 	public String name;
-	public double weight;
+	public int weight;
 	
-	public Keyword(String name,double weight){
+	public Keyword(String name,int weight){
 		this.name = name;
 		this.weight = weight;
 	}
